requirements.txt
python-dotenv
MetaTrader5
pandas
pandas_ta
numpy
